# functions/swagger.yaml
openapi: "3.0.0"
info:
  title: Tech App backend
  description: Documentation for the backend api
  version: 1.1.0
  contact:
    email: unknown39825@gmail.cmo
    url: "https://github.com/technobyte-nitkkr"
servers:
  - url : "http://localhost:5000/techspardha-87928/us-central1/api"
    description: localhost
  - url: "https://us-central1-techspardha-87928.cloudfunctions.net/api"
    description: server
...
components:
  securitySchemes:
    Authdetails:
      type: apiKey
      in: header
      name: Authorization
...
paths:
...
  /contacts:
    get:
      tags:
        - Information
      summary: all technical socities details
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      contacts:
                        type: array
                        items:
                          type: object
                          properties:
                            section:
                              type: string
                            people:
                              type: array
                              items:
                                type: object
                                properties:
                                  imageUrl:
                                    type: string
                                  name:
                                    post: string
        500:
          description: internal server error
  /lectures:
    get:
      tags:
        - Information
      summary: guest lectures api
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      lectures:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                            desc:
                              type: string
                            imageUrl:
                              type: string
                            time:
                              type: string
                            name:
                              type: string
        500:
          description: internal server error
...
  /facts:
    get:
      tags:
        - Information
      summary: about random facts
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      message:
                        type: string
        500:
          description: internal server error
  /videos:
    get:
      tags:
        - Information
      summary: about videos
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        url:
                          type: string
        500:
          description: internal server error
...